pyproject.toml
[tool.poetry]
name = "fuzzylearn"
version = "1.2.0"
description = ""
authors = ["drhosseinjavedani <h.javedani@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.8.5"
wheel = "^0.40.0"
nox = "^2023.4.22"
sklearn = "^0.0.post5"
pandas = "^2.0.2"
numpy = "^1.24.3"
ray = "^2.5.0"
category-encoders = "^2.6.1"
feature-engine = "^1.6.1"
pip = "^23.1.2"
install = "^1.3.5"
python-dotenv = "^1.0.0"
optuna = "^3.2.0"



[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
flake8 = "^6.0.0"
pytest = "^7.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
